Horizontal and Grid listbox layout (#2829)
[eListBox.cpp]
* Add horizontal and grid orientation layouts.
* Optimize rendering to reduce code and improve performance.
* Reformat code to reflect common styling and improve readability.
* add clearBackgroundColor, clearBackgroundColorSelected, clearForegroundColor and clearForegroundColorSelected

[eListBoxContent.cpp]
* Add support for horizontal and grid layouts.
* Optimize rendering to reduce code and improve performance.
* Reformat code to reflect common styling and improve readability.

[TemplatedMultiContent.py]
* Add support for new orientations.

[List.py]
* add enableAutoNavigation, show, hide, goLeft and goRight

[Skin.py]
* Add orientation and item alignment parsers for new horizontal and grid orientation layouts.

[keymap.xml]
* Update "ListboxActions" to separate LEFT/RIGHT for horizontal and grid layouts.

The UI designs that intended to stop using LEFT and RIGHT as the page up/down functions were proposed in the light of this new navigational functionality.  Users are warned that in some cases LEFT and RIGHT need to be reassigned to support these new display layouts.  It is hoped that users will now reconsider their desire to retain LEFT and RIGHT as vertical navigation options as this will allow for the original design of direction buttons becoming more natural and predictable in the actions that they will perform.