SystemInfo.py: LCDClockSetup do not include display type bwlcd128
18:40:48.2151 FileNotFoundError: [Errno 2] No such file or directory: '/usr/share/enigma2/display/clock_skin/clock_lcd_analog.xml'

When the receiver is put on standby (screen StandbySummary) we should not include this type of display. Otherwise the Standby module searches and does not find .../clock_skin/*.xml. According to the setup characteristics of this type of display, Osmio4k and plus do not have this characteristic.
we add this type of siplay in SystemInfo so that it is not included.

solved this problem:
https://openspa.info/threads/bot%C3%B3n-power-no-act%C3%BAa-correctamente.58391/#post-481702