Add video recovery function to 720p
This change gives the possibility of putting the system in video recovery mode, setting standard video at 720p in order to recover the video in case of signal loss on the HDMI port.

The function has two main objectives:
If we change the receiver to a Television with a resolution not supported in the established and saved in the receiver we do not have a signal, before the conventional way was to recover the signal through another program (OpenWebIf) accessing through OSD, now the user can press Long press the EXIT button to access this recovery, displaying a message indicating the instructions to follow and do it all through the remote control and the interface.

The other possible use is to try to recover video through enigma in case the receiver loses it either due to drivers supporting unavailable modes, or due to failures in the HDMI cables used.