[International.py] Add a non ISO country of "KU" for "Kurdistan"
The country entry of "Kurdistan" is being added to address some cultural issues.  The data used for this entry is NOT official.  The values used have been selected to not clash with current ISO designations.  While there is some clash with NATO designations the Enigma2 code is based on ISO so the clashes will be ignored for now.