[MessageBox.py] Restore unused widget definitions
The dummy widget has been restored as some skins do not check if the "list" widget actually exists before manipulating it.  The widget is optional and code should be checking that the widget is available *BEFORE* trying to manipulate it.  This hack has been restored to give skinners more time to correctly fix their skins!