LocaleSelection.py: do not disable the selected language variable
The text is translated when we select current locale according to the available packages. If it is selected and the GREEN button is pressed, the first message to restart the GUI is already translated into the language that is intended to be changed.

This change does allow us to change the language and thus change the OSD.