ChannelSelection.py: add new funtion create bouquets from Last Scanned
Last Scanned is unique, renaming it or not is indifferent, we always depend on the last scan. Before renaming this bouquet there is no value since we always depend on the last scan. With this change, renaming it means creating an independent and unlinked bouquet through the last scan, in this way we can create as many bouquets as we want with Last Scanned.

* I have corrected an indentation in another block of code.

* Last Scanned becomes a translated bouquet