[OSDCalibration.py] Prepare for a major refactor
This change is in preparation of a major refactor of the "UserInterfacePositioner.py" code and its associated Wizard "WizardUserInterfacePositioner.py".  This commit has minimal code changes and is focused on introducing the new module names and new classes.  This will give skinners time to prepare skins for the code changes when they are committed.

At this time skinners should be aware that the new screen name for "UserInterfacePositioner2" will be "OSDCalibration".  When this commit is merged skinners should add a conditional widget "description" that will replace the current widget "status".  The "conditional" attribute should be used on the "description" widget to suppress skin errors until the new OSDCalibration code based on "Setup" is published.  Skinners should also add the following widget to their skins:
        <widget name="footnote" position="0,0" size="0,0" conditional="footnote" />

The "status" widget should also be made "conditional" to suppress skin errors when the new OSDCalibration code based on "Setup" is published.  The new "description" widget is identically functional to the "description" widget in "Setup" based screens.  This change is required because the new "OSDCalibration" screen will be a sub-class of the current "Setup" class.  When the new code is activated the "status" widget and the conditional="footnote" (from the line added above) can be removed from the "OSDCalibration" screen.